Added a game scene
The game now features a main menu scene and a game scene. When the player clicks the "Play Game" button in the main menu the game scene is loaded. The game scene contains a player which can fly around the world. When the player presses the quit game button on the main menu, the window closes.