Finished the game over menu functionality
I added a game over menu to the game scene with placeholder images. When the player the required number of enemies the game over is triggered and the menu appears. The player can click the play again button to play again or the main menu button to go back to the main menu. Currently, to increment number of enemies killed for testing you can press tab to simulate the game over.