Added PlayerHUD and Shooting
I added the player lives to the player hud and the number of enemies remaining in the wave. I also added shooting to the game. The player can now press space bar to shoot a bullet. The bullet spawns at the player location and rotation and moves in the direction of it's rotation upon spawning. If the bullet goes off screen, the bullet is inactivated and moved back to the bullet pool location. If the bullet hasn't been deactivated for more than 10 seconds the bullet is deactivated and moved to the bullet pool.